refactor(bulk-upload): Break down validate endpoint
